pyproject.toml
[project]
name = "PaperQuant"
version = "0.1.0"
description = "An interactive & open-source paper trading platform for algorithmic traders."
authors = [
    {name = "DikshitRJ",email = "ourledtv@gmail.com"}
]
license = "GPL-3.0-or-later"
license-files = ["LICENSE"]
readme = "README.md"
requires-python = "3.10.19"
dependencies = [
    "yfinance (>=1.1.0,<2.0.0)",
    "jupyter (>=1.1.1,<2.0.0)",
    "diskcache (>=5.6.3,<6.0.0)",
    "pyzmq (>=27.1.0,<28.0.0)"
]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

